python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
rembg==2.0.50
rapidocr-onnxruntime==1.3.15
